refactor: WIP for multi-backend
Most of interfaces are modified, for multi-backend

However, there's some breaking changes that blocks version control over
sandboxs, thus WIP.

This refactor is staged until a datapath doc is available, and also with
a further (and complete) design for this refactor. Refer to the
`docs/datapath` for progress.

Since it's broken, it won't be directly merged into the dev branch until
completed.